2.0.2
Replaced printf_s() with printf() in sutest.c for compatibility on Linux.
Replaced check for NDEBUG with a check for Windows, in which case DebugBreak() is called.

2.0.0

Changed license to Apache Version 2.0

2.0.1

Added comment to ring_buf.h about assumptions for "atomicity" of the RingBufCtr type.

2.0.1

2.0.1

added 'volatile' to the head and tail indexes into the ring buffer